Put class before href in link replacements
The links were not being replaced correctly, because a regex was looking for the href to come before the classes on each link. The renderer creates links with attributes in the other order: classes, then links. This change fixes the link replacement behavior to correctly patch generated links on page create, move, and delete.