Change most formatting to wrapOrInsert
Like in most editors, it can make sense to start formatting before you've written the text to format. This change allows the format buttons and existing shortcuts to apply that behavior by inserting their respective start marker when there is no selection. With a selection, they behave as before and wrap it in the correct start/end markers.