Use current path for new page dialog
The new page dialog will now suggest a path that creates a subpage of the current page when the button is clicked. The reasoning is that it's much easier to erase the path when you want to make a root page than it is to type out the path or search for it in the directory listing when you want to make a subpage.